[MDAPI-423][C++] Fixed Java-to-C++ porting errors in packed event fields, native conversions, string handling, and candle attributes. (#96)
* `setSequence()` and `OrderBase::setSource()` now preserve unrelated
packed bits; `Message` native conversion
preserves event time; `StringLike` copy/move operations and native
C-string handling are now safe.
* Candle equality, formatting, price-level validation/parsing, and
session parsing now match Java QD semantics.
Price levels reject negative values, negative zero, infinities, and
trailing parse data; floating-point equality
follows Java NaN/signed-zero rules; only case-insensitive `"true"`
enables the regular candle session.
* APIs that can fail validation or parsing are no longer `noexcept`;
invalid input now throws an exception instead
    of terminating the process.

+++ src/event/market/TradeBase.cpp
@@ -120,7 +120,8 @@ void TradeBase::setSequence(std::int32_t sequence) {
         throw InvalidArgumentException("Invalid sequence value = " + std::to_string(sequence));
     }
 
-    tradeBaseData_.timeSequence = orOp(andOp(tradeBaseData_.timeSequence, ~MAX_SEQUENCE), sequence);
+    tradeBaseData_.timeSequence =
+        orOp(andOp(tradeBaseData_.timeSequence, ~static_cast<std::int64_t>(MAX_SEQUENCE)), sequence);
 }
 
 std::int16_t TradeBase::getExchangeCode() const noexcept {
